Move log collection in post-run zuul step
Currently 99-logs.yml playbook is used to collect logs
during edpm deployment and runned as a part of
deploy-edpm.yml playbook. But any playbook before 99-logs.yml
playbook fails in deploy-edpm.yaml, then log collection fails.

In order to fix this issue, we are going to call 99-logs.yml
from e2e-collect-logs.yml playbook. It will be called as a part
of post-run zuul step so that log collection always happen as.

It will be runned in e2e and edpm jobs.

Note: it drops edpm/post.yml playbook (which is not working as
expected.)

Signed-off-by: Chandan Kumar <raukadah@gmail.com>